Shorten processor hierarchy
